Store InstrumentationFlags and Operators as separate vecs
Instead of storing function bodies as a `Vec` of pairs of `Operator` and
`InstrumentationFlag` (in the `Instruction` type), we keep them in
separate vectors with the `InstrumentationFlag` vector optional. This
allows faster parsing and encoding for use cases that don't use
`InstrumentationFlag`. For cases that do use it, the additional vector
will be allocated as needed only for the functions which use it. This
results in as much as a 50% speedup in parsing and encoding times when
`InstrumentationFlag` isn't used.